requirements.txt
akshare==1.18.30
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
baostock==0.8.9
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.14.0
cycler==0.12.1
decorator==5.2.1
et_xmlfile==2.0.0
fastapi==0.135.1
fonttools==4.61.1
greenlet==3.3.2
h11==0.16.0
html5lib==1.1
idna==3.11
jsonpath==0.82.2
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
mini-racer==0.14.1
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
tabulate==0.9.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
webencodings==0.5.1
xlrd==2.0.2
